requirements.txt
blinker==1.8.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
dominate==2.9.1
Flask==3.0.3
Flask-Bootstrap4==4.0.2
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pycparser==2.22
requests==2.32.3
urllib3==2.2.2
visitor==0.1.3
Werkzeug==3.0.3
opentelemetry-api                        ==1.35.0
opentelemetry-exporter-jaeger            ==1.21.0
opentelemetry-exporter-jaeger-proto-grpc ==1.21.0
opentelemetry-exporter-jaeger-thrift     ==1.21.0
opentelemetry-exporter-otlp              ==1.35.0
opentelemetry-exporter-otlp-proto-common ==1.35.0
opentelemetry-exporter-otlp-proto-grpc   ==1.35.0
opentelemetry-exporter-otlp-proto-http   ==1.35.0
opentelemetry-instrumentation            ==0.56b0
opentelemetry-instrumentation-flask      ==0.56b0
opentelemetry-instrumentation-requests   ==0.56b0
opentelemetry-instrumentation-wsgi       ==0.56b0
opentelemetry-proto                      ==1.35.0
opentelemetry-sdk                        ==1.35.0
opentelemetry-semantic-conventions       ==0.56b0
opentelemetry-util-http                  ==0.56b0
gunicorn
